# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Liketexts API
  version: 1.0.0
  description: 'Substack-style platform: readers subscribe to writers via allocation
    slots; posts are paywalled (preview + subscriber content).'
servers:
- url: http://localhost:5000
paths:
...
  /api/v1/posts:
    post:
      summary: Create a draft or scheduled post
      tags:
      - Publishing
      description: Any authenticated user may create a first post — doing so grants
        the WRITER capability.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedPost'
        '400':
          description: Missing title/preview or bad scheduled_at.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
  /api/v1/posts/{post_id}/publish:
    post:
      summary: Publish a post immediately
      tags:
      - Publishing
      parameters:
      - name: post_id
        in: path
        required: true
        description: Draft post id.
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Published post.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishedPost'
        '400':
          description: Invalid post_id format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
  /api/v1/posts/{post_id}/schedule:
    post:
      summary: Schedule a draft post
      tags:
      - Publishing
      parameters:
      - name: post_id
        in: path
        required: true
        description: Draft post id.
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - scheduled_at
              properties:
                scheduled_at:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Scheduled post.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledPost'
        '400':
          description: Missing/invalid scheduled_at or post_id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
  /api/v1/posts/{post_id}:
    delete:
      summary: Cancel a scheduled post
      tags:
      - Publishing
      parameters:
      - name: post_id
        in: path
        required: true
        description: Scheduled post id.
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Cancelled post.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelledPost'
        '400':
          description: Invalid post_id format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
    patch:
      summary: Update post content
      tags:
      - Publishing
      parameters:
      - name: post_id
        in: path
        required: true
        description: Post id.
        schema:
          type: string
          format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Updated post.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedPost'
        '400':
          description: Invalid post_id format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
    get:
      summary: Get a single post with paywall logic
      tags:
      - Publishing
      description: Public preview; no auth required.
      parameters:
      - name: post_id
        in: path
        required: true
        description: Post id.
        schema:
          type: string
          format: uuid
      - *id001
      responses:
        '200':
          description: Paywall-applied post.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostView'
        '400':
          description: Invalid post_id format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/posts/recent:
    get:
      summary: Latest published posts
      tags:
      - Publishing
      description: Public homepage feed, always preview-masked.
      parameters:
      - name: limit
        in: query
        required: false
        description: Max posts to return.
        schema:
          type: integer
          default: 10
          maximum: 50
      responses:
        '200':
          description: Recent posts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentPosts'
  /api/v1/posts/feed:
    get:
      summary: Reader feed from allocated writers
      tags:
      - Publishing
      description: Empty without an active subscription/allocation (archival included).
      parameters:
      - name: limit
        in: query
        required: false
        description: Max posts to return.
        schema:
          type: integer
          default: 20
      - name: cursor
        in: query
        required: false
        description: Opaque pagination cursor (published_at of the last seen post).
        schema:
          type: string
      responses:
        '200':
          description: Feed page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
        '401':
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
  /api/v1/posts/writer:
    get:
      summary: Current writer's posts
      tags:
      - Publishing
      parameters:
      - name: status
        in: query
        required: false
        description: Filter by status.
        schema:
          type: string
          enum:
          - draft
          - scheduled
          - published
          - cancelled
      responses:
        '200':
          description: Own posts, unmasked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WriterPosts'
        '401':
          description: Missing, invalid, or expired access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
...
components:
  schemas:
    Error:
      type: object
      required:
      - title
      - status
      - detail
      - message
      - code
      properties:
        title:
          type: string
          example: Unauthorized
        status:
          type: integer
          example: 401
        detail:
          type: string
        message:
          type: string
          description: Alias of detail.
        code:
          type: string
          example: UNAUTHORIZED
...
    PostView:
      type: object
      description: Paywall-applied post. `subscriber_content` is null without an active
        allocation to the writer (or authorship).
      required:
      - id
      - writer_id
      - title
      - preview_content
      - status
      - published_at
      - created_at
      - subscriber_content
      - has_full_access
      - writer_name
      - writer_avatar_url
      properties:
        id:
          type: string
          format: uuid
        writer_id:
          type: string
          format: uuid
        title:
          type: string
        preview_content:
          type: string
        status:
          type: string
          example: published
        published_at: &id003
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        subscriber_content: *id002
        has_full_access:
          type: boolean
        writer_name: *id002
        writer_avatar_url: *id002
    CreatePostRequest:
      type: object
      required:
      - title
      - preview_content
      properties:
        title:
          type: string
        preview_content:
          type: string
        subscriber_content:
          type: string
          default: ''
        scheduled_at:
          type: string
          format: date-time
          description: ISO 8601. When present the post is created scheduled.
    CreatedPost:
      type: object
      required:
      - id
      - title
      - status
      - scheduled_for
      - created_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        scheduled_for: *id003
        created_at:
          type: string
          format: date-time
    PublishedPost:
      type: object
      required:
      - id
      - title
      - status
      - published_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        published_at: *id003
    ScheduledPost:
      type: object
      required:
      - id
      - title
      - status
      - scheduled_for
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        scheduled_for: *id003
    CancelledPost:
      type: object
      required:
      - id
      - title
      - status
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
    UpdatedPost:
      type: object
      required:
      - id
      - title
      - status
      - updated_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        updated_at:
          type: string
          format: date-time
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
        preview_content:
          type: string
        subscriber_content:
          type: string
    Feed:
      type: object
      required:
      - posts
      - next_cursor
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostView'
        next_cursor: *id002
    RecentPosts:
      type: object
      description: Latest published posts, always preview-masked (public).
      required:
      - posts
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostView'
    WriterPost:
      type: object
      description: Full post for the owning writer (no paywall masking).
      required:
      - id
      - title
      - status
      - preview_content
      - subscriber_content
      - scheduled_for
      - published_at
      - created_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        preview_content:
          type: string
        subscriber_content:
          type: string
        scheduled_for: *id003
        published_at: *id003
        created_at:
          type: string
          format: date-time
    WriterPosts:
      type: object
      required:
      - posts
      - subscriber_count
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/WriterPost'
        subscriber_count:
          type: integer
          description: Readers with an active allocation to this writer.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/v1/auth/login.